baserlib: Add more generic base class for opcode resources
This base class, which employs a meta class for the actual work, allows
easier definitions of RAPI resources using opcodes. Follow-up patches
will change some of the existing RAPI resources.

The long-term goal with these changes is to make it easier to verify the
consistency of RAPI with CLI (or the underlying opcodes).

Signed-off-by: Michael Hanselmann <hansmi@google.com>
Reviewed-by: René Nussbaumer <rn@google.com>